add notification

+++ src/main/java/com/tytao/community/dto/CommentDTO.java
@@ -12,7 +12,7 @@ public class CommentDTO {
     private Long gmtCreate;
     private Long gmtModified;
     private Long likeCount;
-    private Long commentCount;
+    private Integer commentCount;
     private String content;
     private User user;
 }
